Refactored HTML scraping
